+++ Aquaria/GridRender.cpp
@@ -122,7 +122,7 @@ SongLineRender::SongLineRender()
 
 void SongLineRender::newPoint(const Vector &pt, const Vector &color)
 {
-	int maxx = 40;
+	size_t maxx = 40;
 	bool inRange = true;
 	if (pts.size() > 1)
 		inRange = (pt - pts[pts.size()-2].pt).isLength2DIn(4);
@@ -155,10 +155,10 @@ void SongLineRender::onRender()
 	if (ls < 0)
 		ls = 1;
 	glLineWidth(ls);
-	const int alphaLine = pts.size()*(0.9f);
+	const unsigned int alphaLine = pts.size()*(0.9f);
 	float a = 1;
 	glBegin(GL_LINE_STRIP);
-	for (int i = 0; i < pts.size(); i++)
+	for (size_t i = 0; i < pts.size(); i++)
 	{
 		if (i < alphaLine)
 			a = float(i)/float(alphaLine);
